Build list of modules from sanitized directory listing. Add script definition in package.json to build rules.
Auditors: @mrose17
