+++ tools/build-rules.js
@@ -1,4 +1,4 @@
-const rules = require('../rules')
+const rules = require('../categories')
 
 rules.all(function (err, definition) {
   console.log(JSON.stringify(definition, null, 2))
